playbook
修正

+++ ansible/copy_zsh_setting_file.sh
@@ -2,6 +2,8 @@
 
 for file in /home/vagrant/.zprezto/runcoms/z*
 do
-	ln -s ${file} /home/vagrant/.${file##*/}
+	if [ ! -e /home/vagrant/.${file##*/} ]; then
+		ln -s ${file} /home/vagrant/.${file##*/}
+	fi
 done
